Extract content module loading to a separate class
