+++ intellij-plugin-structure/structure-intellij/src/main/java/com/jetbrains/plugin/structure/intellij/plugin/IdePluginManager.kt
@@ -33,13 +33,10 @@ import com.jetbrains.plugin.structure.base.utils.toSystemIndependentName
 import com.jetbrains.plugin.structure.base.utils.withPathSeparatorOf
 import com.jetbrains.plugin.structure.intellij.extractor.ExtractorResult
 import com.jetbrains.plugin.structure.intellij.extractor.PluginExtractor.extractPlugin
-import com.jetbrains.plugin.structure.intellij.plugin.Module.FileBasedModule
-import com.jetbrains.plugin.structure.intellij.plugin.Module.InlineModule
 import com.jetbrains.plugin.structure.intellij.plugin.PluginCreator.Companion.createInvalidPlugin
 import com.jetbrains.plugin.structure.intellij.plugin.PluginCreator.Companion.createPlugin
-import com.jetbrains.plugin.structure.intellij.plugin.descriptors.DescriptorResource
+import com.jetbrains.plugin.structure.intellij.plugin.loaders.ContentModuleLoader
 import com.jetbrains.plugin.structure.intellij.plugin.module.ContentModuleScanner
-import com.jetbrains.plugin.structure.intellij.problems.AnyProblemToWarningPluginCreationResultResolver
 import com.jetbrains.plugin.structure.intellij.problems.IntelliJPluginCreationResultResolver
 import com.jetbrains.plugin.structure.intellij.problems.PluginCreationResultResolver
 import com.jetbrains.plugin.structure.intellij.problems.PluginLibDirectoryIsEmpty
@@ -59,7 +56,6 @@ import org.jdom2.input.JDOMParseException
 import org.slf4j.LoggerFactory
 import java.io.File
 import java.io.IOException
-import java.net.URI
 import java.nio.file.Files
 import java.nio.file.Path
 import java.time.Duration
@@ -79,6 +75,8 @@ class IdePluginManager private constructor(
   private val THIRD_PARTY_LIBRARIES_FILE_NAME = "dependencies.json"
 
   private val optionalDependencyResolver = OptionalDependencyResolver(this::loadPluginInfoFromJarOrDirectory)
+
+  private val contentModuleLoader = ContentModuleLoader(this::loadPluginInfoFromJarOrDirectory)
 
   private val contentModuleScanner = ContentModuleScanner(fileSystemProvider)
 
@@ -155,35 +153,6 @@ class IdePluginManager private constructor(
     } catch (e: JarArchiveCannotBeOpenException) {
       LOG.warn("Unable to extract {} (searching for {}): {}", jarFile, descriptorPath, e.getShortExceptionMessage())
       createInvalidPlugin(jarFile, descriptorPath, UnableToExtractZip())
-    }
-  }
-
-  private fun loadModuleFromDescriptorResource(
-    moduleId: String,
-    descriptorResource: DescriptorResource,
-    parentPlugin: PluginCreator? = null,
-    resourceResolver: ResourceResolver
-  ): PluginCreator {
-    return descriptorResource.inputStream.use {
-      try {
-        val problemResolver = AnyProblemToWarningPluginCreationResultResolver
-        val descriptorXml = JDOMUtil.loadDocument(it)
-        createPlugin(
-          descriptorResource,
-          parentPlugin,
-          descriptorXml,
-          resourceResolver,
-          problemResolver
-        ).also {
-          logPluginCreationWarnings(moduleId, it)
-        }
-      } catch (e: IOException) {
-        with(descriptorResource) {
-          LOG.warn("Unable to read descriptor stream (source: '$uri')", e)
-          val problem = UnableToReadDescriptor(fileName, e.localizedMessage)
-          createInvalidPlugin(artifactFileName, fileName, problem)
-        }
-      }
     }
   }
 
@@ -359,32 +328,7 @@ class IdePluginManager private constructor(
   }
 
   private fun resolveContentModules(pluginFile: Path, currentPlugin: PluginCreator, resourceResolver: ResourceResolver, problemResolver: PluginCreationResultResolver) {
-    if (currentPlugin.isSuccess) {
-      val contentModules = currentPlugin.plugin.contentModules
-      for (module in contentModules) {
-        when (module) {
-          is FileBasedModule -> {
-            val configFile = module.configFile
-            val moduleCreator = loadPluginInfoFromJarOrDirectory(
-              pluginFile,
-              configFile,
-              false,
-              resourceResolver,
-              currentPlugin,
-              problemResolver
-            )
-            currentPlugin.addModuleDescriptor(module.name, module.loadingRule, configFile, moduleCreator)
-          }
-
-          is InlineModule -> {
-            val moduleDescriptorResource = getDescriptorResource(module, pluginFile, currentPlugin.descriptorPath)
-            val moduleCreator =
-              loadModuleFromDescriptorResource(module.name, moduleDescriptorResource, currentPlugin, resourceResolver)
-            currentPlugin.addModuleDescriptor(module, module.loadingRule, moduleDescriptorResource, moduleCreator)
-          }
-        }
-      }
-    }
+    contentModuleLoader.resolveContentModules(pluginFile, currentPlugin, resourceResolver, problemResolver)
   }
 
   private fun extractZipAndCreatePlugin(
@@ -502,27 +446,6 @@ class IdePluginManager private constructor(
     with(telemetry) {
       parsingDuration = Duration.ofMillis(pluginCreationDurationInMillis)
       archiveFileSize = pluginFile.pluginSize
-    }
-  }
-
-  private fun getDescriptorResource(module: InlineModule, pluginFile: Path, descriptorPath: String): DescriptorResource {
-    // TODO descriptor path is not relative to the pluginFile JAR. See MP-7224
-    val parentUriStr = if (pluginFile.isJar()) {
-      "jar:" + pluginFile.toUri().toString() + "!" + descriptorPath.toSystemIndependentName()
-    } else {
-      pluginFile.toUri().toString() + "/" + descriptorPath.toSystemIndependentName()
-    }
-    val uriStr = parentUriStr + "#modules/" + module.name
-    return DescriptorResource(module.textContent.byteInputStream(), URI(uriStr), URI(parentUriStr))
-  }
-
-  private fun logPluginCreationWarnings(pluginId: String, pluginCreator: PluginCreator) {
-    val pluginCreationResult = pluginCreator.pluginCreationResult
-    if (LOG.isDebugEnabled && pluginCreationResult is PluginCreationSuccess) {
-      val warningMessage = pluginCreationResult.warnings.joinToString("\n") {
-        it.message
-      }
-      LOG.debug("Plugin or module '$pluginId' has plugin problems: $warningMessage")
     }
   }
